+++ commons-mail/src/main/java/com/hb0730/commons/mail/spring/service/MailService.java
@@ -43,4 +43,9 @@ public interface MailService {
      * 测试连接
      */
     void testConnection();
+
+    /**
+     * 清除缓存
+     */
+    void clearCache();
 }